requirements.txt
asgiref==3.5.2 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.1.0 ; python_version >= "3.8" and python_version < "4.0"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
django-cors-headers==3.13.0 ; python_version >= "3.8" and python_version < "4.0"
django==4.1.2 ; python_version >= "3.8" and python_version < "4.0"
djangorestframework-simplejwt==5.2.2 ; python_version >= "3.8" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.8" and python_version < "4.0"
drf-spectacular==0.24.2 ; python_version >= "3.8" and python_version < "4.0"
importlib-resources==5.10.0 ; python_version >= "3.8" and python_version < "3.9"
inflection==0.5.1 ; python_version >= "3.8" and python_version < "4.0"
jsonschema==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
pyjwt==2.6.0 ; python_version >= "3.8" and python_version < "4.0"
pyrsistent==0.19.1 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.5 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.8" and python_version < "4.0"
tzdata==2022.5 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.10.0 ; python_version >= "3.8" and python_version < "3.9"
